Renamed highlighter package to structure.
