+++ src/main/java/org/netbeans/asciidoc/structure/AsciidoctorFoldTypeProvider.java
@@ -1,4 +1,4 @@
-package org.netbeans.asciidoc.highlighter;
+package org.netbeans.asciidoc.structure;
 
 import java.util.Arrays;
 import java.util.Collection;
