+++ src/main/java/org/netbeans/asciidoc/structure/AsciidoctorStructureScanner.java
@@ -1,4 +1,4 @@
-package org.netbeans.asciidoc.highlighter;
+package org.netbeans.asciidoc.structure;
 
 import java.util.ArrayDeque;
 import java.util.ArrayList;
